Giving iterators their own reserved connection.
It seems to dangerous to let iterators silently obtain an existing
connection. The external connection might be used in a way that
is not compatible with the iterator and outer and inner calls (from
the view of the iterator) might be intertwined, having side effects.
Thus, iterators reserve their own connection.