Fixed a memory leak associated with connection reservation.
Reservations were recorded with a map. Thus, each Thread that had
reserved a connection existed as key in this map. Since the entries
of the map were never evicted, a memory leak occurred when a lot
of different threads would reserve connections. This has been solved
by using a guava cache with weak keys (those get evicted by the
garbage collector when there are no other references to the key).