Update methods now require specification about mirror subset resets.
Until now, each update of a data table caused its mirror subsets
to reset the updated rows. However, since the annotations are stored
with in the data table directly in newer versions of JeDIS, this
is a major bug causing mirrors to reset rows which just had been
processed. To avoid this, the new parameter was introduced which
can be set to false when required in the XMI DB Writer.