Fixing small connection chaos.
The new postgres JDBC driver caused an issue with the LOCK TABLE directive on subset reset because it requires to be part of an explicit transaction block started with `BEGIN TRANSACTION`. Changes here caused issues with the multiple re-use of connections.
There are now connections that are not open for sharing through a new parameter when reserving connections. All tests running.