Fix a bug causing a PK insertion clash into mirror subsets.
It could happen that is was attempted to add an entry already present in a mirror subset again. When there were multiple mirror subsets and some had the inserted row and others hadn't, the row would go into the first mirror subset in the list of mirror subsets instead of the mirror subset actually missing the entry.