Un-synchronize connection closing.
There were deadlocks were threads wanted to close their connections but waited on the monitor held by oder threads waiting for a connection. Now, collections can be closed at any time. This might cause closed connection to be obtained so we put a loop into obtainConnection() that checks if the connection is closed and, if so, tries again.